Keep the info button when clicking on the legend
- Need to redraw the info tooltip when the the entire stacked bar
graph is redrawn once the user filters for a particular food group
