+++ src/ui/filter.js
@@ -1001,6 +1001,71 @@ export function clearSandbox() {
   displayOverrideFoodFilter();
 }
 
+
+// infoIconHover(element): When the info icon is being hovered over
+export function infoIconOnHover(element) {
+    element = d3.select(element);
+    element.classed("infoIcon-enabled", true);
+    element.classed("infoIcon-disabled", false);
+}
+
+// infoIconUnHover(element): when the info icon is unhovered
+export function infoIconUnHover(element) {
+    element = d3.select(element);
+
+    // do not change back the color if the icon has already been clicked
+    if (element.attr("infoIconClicked") == null) {
+        element.classed("infoIcon-enabled", false);
+        element.classed("infoIcon-disabled", true);
+    }
+}
+
+// drawToolTips(toolTipId, elementsWithInfoIcons, toolTipTextFunc): Draws the tooltips
+export function drawToolTips(toolTipId, elementsWithInfoIcons, toolTipTextFunc = undefined) {
+    // ----------- draw the tool tips ---------------
+    
+    const infoIconGroups = elementsWithInfoIcons.append("span");
+    const icons = infoIconGroups.append("i")
+        .attr("class", "fa fa-info-circle infoIcon")
+        .attr("aria-hidden", true); // used for accessibility purposes
+
+    if (toolTipTextFunc == undefined) {
+        toolTipTextFunc = (data) => { return data; };
+    }
+
+    icons.attr("id", toolTipId)
+        .attr("title", toolTipTextFunc)
+        .attr("data-bs-html", "true")
+        .attr("data-toggle", "tooltip")
+        .attr("data-placement", "right")
+        .each((data, index, elements) => { $(elements[index]).tooltip({placement: "right", container: "body", trigger: "manual"}); })
+
+        // rewrite the title again since creating a Bootstrap tooltip will set the 'title' attribute to null
+        //   and transfer the content of the 'title' attribute to a new attribute called 'data-bs-original-title'
+        //
+        // Comment out the line below if we want to add back the title attribute.
+        // Its used for the hover text of the icon, but the user can see the same text if they click the icon
+        //
+        // The 'title' attribute seems to be for some assessbility purposes
+        // https://fontawesome.com/v5/docs/web/other-topics/accessibility
+        //
+        // .attr("title", toolTipTextFunc);
+
+    // add the hidden text needed for screen readers
+    for (const element of icons._groups[0]) {
+        d3.select(element.parentNode).append("span")
+            .classed("sr-only", true)
+            .text(toolTipTextFunc);
+    }
+
+    icons.on("mouseenter", (event) => { 
+        infoIconOnHover(event.target)
+    });
+    icons.on("mouseleave", (event) => { infoIconUnHover(event.target)});
+
+    // ----------------------------------------------
+}
+
 /**
  * A function that retrieves the current user-selected filters, and filters the contaminant data from the global
  * TDS data for the currently-selected chemical group, chemical, and years. This function is called before passing
